requirements.txt
Django>=3.2.0,<4.0.0
django-cors-headers>=3.10.0
djangorestframework>=3.12.0
Pillow>=8.2.0
python-dotenv>=0.19.0 